// front/src/types/user.ts
type User = {
  client_id: string | null;
  connection: string | null;
  email: string | null;
  email_verified: boolean | null;
  form: {} | null;
  is_signup: boolean | null;
  request_language: string | null;
  tenant: string | null;
  transaction: {} | null;
  usePassKey: boolean | null;
  username: string | null;
  _id: string | null;
};

export default User;